Move stack iteration test into its own file
This prevents other tests from interfering with the js engine
initialization
